requirements.txt
aiofiles==25.1.0
aioredis==2.0.1
aiosmtplib==5.0.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
apturl==0.5.2
async-timeout==5.0.1
attrs==21.2.0
babel==2.17.0
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==6.2.4
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==46.0.3
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.8.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.121.3
fastapi-admin==1.0.4
fastapi-mail==1.6.1
fasteners==0.14.1
future==0.18.2
google-auth==2.45.0
greenlet==3.3.0
gyp==0.1
h11==0.16.0
httplib2==0.20.2
httptools==0.7.1
idna==3.3
importlib-metadata==4.6.4
iso8601==2.1.0
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
motor==3.7.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
passlib==1.7.4
pendulum==3.1.0
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.15.4
PyNaCl==1.5.0
pyparsing==2.4.7
pypika-tortoise==0.6.3
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2025.11.3
reportlab==3.6.8
requests==2.25.1
rsa==4.9.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.45
ssh-import-id==5.11
starlette==0.50.0
systemd-python==234
texttable==1.6.4
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.38.0
uvloop==0.22.1
wadllib==1.3.6
watchfiles==1.1.1
websocket-client==1.2.3
websockets==15.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
